Now UDOO Image gets created

+++ scripts/udooimage.sh
@@ -37,21 +37,31 @@ done
 IMG_FILE="Volumio.img"
  
 echo "Creating Image Bed"
-dd if=/dev/zero of=${IMG_FILE} bs=1M count=1000
+dd if=/dev/zero of=${IMG_FILE} bs=1M count=1048
 LOOP_DEV=`sudo losetup -f --show ${IMG_FILE}`
  
 sudo parted -s "${LOOP_DEV}" mklabel msdos
-sudo parted -s "${LOOP_DEV}" unit cyl mkpart primary ext3 -- 2cyl -0
+#sudo parted -s "${LOOP_DEV}p1" mkpart primary ext3 0 10
+sudo parted -s "${LOOP_DEV}" mkpart primary ext3 10 1048
 sudo parted -s "${LOOP_DEV}" set 1 boot on
 sudo parted -s "${LOOP_DEV}" print
 sudo partprobe "${LOOP_DEV}"
+sudo kpartx -a "${LOOP_DEV}"
  
+LOOP_PART=`echo /dev/mapper/"$( echo $LOOP_DEV | sed -e 's/.*\/\(\w*\)/\1/' )"p1`
+
+if [ ! -b "$LOOP_PART" ]
+then
+	echo "$LOOP_PART doesn't exist"
+	exit 1
+fi
+
 echo "Creating filesystems"
-sudo mkfs.ext4 -O ^has_journal -E stride=2,stripe-width=1024 -b 4096 "${LOOP_DEV}" -L volumio
+sudo mkfs.ext4 -E stride=2,stripe-width=1024 -b 4096 "${LOOP_PART}" -L volumio
 sync
  
 echo "Burning bootloader"
-sudo dd if=/dev/zero of=${LOOP_DEV} bs=1k count=1023 seek=1
+#sudo dd if=/dev/zero of=${LOOP_DEV} bs=1k count=1023 seek=1
 sudo dd if=platforms/udoo/uboot/u-boot-q.imx of=${LOOP_DEV} bs=512 seek=2
 sync
 
@@ -59,7 +69,7 @@ sync
 echo "Copying Volumio RootFs"
 sudo mkdir /mnt
 sudo mkdir /mnt/volumio
-sudo mount -t ext4 "${LOOP_DEV}" /mnt/volumio
+sudo mount -t ext4 "${LOOP_PART}" /mnt/volumio
 sudo rm -rf /mnt/volumio/*
 sudo cp -r build/root/* /mnt/volumio
 fi
